Added Activities many things added to Char screen
Need to work on the Char screen and make it nice and pretty and then
make it work. Figure out how I'm going to use the Compendium screen to
send me to the Arcana Screen and have the Arcana Screen only have the
certain ammount of buttons it needs depending on the Trump chosen on the
Compendium screen. From there it needs to start the CharScreen with the
right values depending on what was chosen on the Arcana screen.