Added some to the Char screen
Have the base for where the numerical values will go. Need to work on
the arcans screen make an array for it and have it set so that 0-21 sets
the content view to what it needs to be. Need to make a bundle on the
compendium, make a trump screen have maybe something like
{ Bundle
setkey?("trump")
("Trump""Fool")
}